Extract common methods to Utils
