+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/marker/PropertiesResourceConnectionLineMarkerProvider.java
@@ -7,12 +7,11 @@ package com.microsoft.azure.toolkit.intellij.connector.marker;
 
 import com.intellij.lang.properties.psi.Property;
 import com.intellij.lang.properties.psi.impl.PropertyValueImpl;
-import com.intellij.openapi.module.Module;
-import com.intellij.openapi.module.ModuleUtil;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.microsoft.azure.toolkit.intellij.connector.Connection;
 import com.microsoft.azure.toolkit.intellij.connector.ResourceDefinition;
+import com.microsoft.azure.toolkit.intellij.connector.Utils;
 import com.microsoft.azure.toolkit.intellij.connector.spring.SpringSupported;
 import com.microsoft.azure.toolkit.lib.common.model.AzResource;
 import org.apache.commons.collections4.CollectionUtils;
@@ -24,7 +23,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.regex.Matcher;
 
 public class PropertiesResourceConnectionLineMarkerProvider extends AbstractResourceConnectionLineMarkerProvider {
     @Override
@@ -34,23 +32,15 @@ public class PropertiesResourceConnectionLineMarkerProvider extends AbstractReso
 
     @Override
     protected Connection<? extends AzResource, ?> getConnectionForPsiElement(@Nonnull PsiElement element) {
-        final String value = getEnvironmentVariableFromPsiElement(element);
-        final Module module = ModuleUtil.findModuleForPsiElement(element);
-        final Connection<AzResource, ?> connection = AbstractResourceConnectionLineMarkerProvider.getConnectionWithEnvironmentVariable(module, value);
+        final Connection<? extends AzResource, ?> connection = Utils.getConnectionForPropertiesValue((PropertyValueImpl) element);
         if (Objects.isNull(connection)) {
             return null;
         }
-        final ResourceDefinition<AzResource> definition = connection.getResource().getDefinition();
-        final List<Pair<String, String>> variables = definition instanceof SpringSupported<AzResource> ?
-                ((SpringSupported<AzResource>) definition).getSpringProperties() : Collections.emptyList();
+        final ResourceDefinition<? extends AzResource> definition = connection.getResource().getDefinition();
+        final List<Pair<String, String>> variables = definition instanceof SpringSupported<? extends AzResource> ?
+                ((SpringSupported<? extends AzResource>) definition).getSpringProperties() : Collections.emptyList();
         final Property parent = PsiTreeUtil.getParentOfType(element, Property.class);
         final String property = Optional.ofNullable(parent).map(Property::getKey).orElse(null);
         return CollectionUtils.isNotEmpty(variables) && StringUtils.equalsIgnoreCase(variables.get(0).getKey(), property) ? connection : null;
-    }
-
-    private String getEnvironmentVariableFromPsiElement(@Nonnull PsiElement element) {
-        final String text = element.getText();
-        final Matcher matcher = SPRING_PROPERTY_VALUE_PATTERN.matcher(text);
-        return matcher.matches() ? matcher.group(1) : text;
     }
 }
